src/services/nlp_service/requirements.txt
uvicorn
fastapi
google-generativeai
python-dotenv